backend/Process-interview/requirements.txt
# absl-py==1.0.0
# appdirs==1.4.4
# astunparse==1.6.3
# audioread==2.1.9
# cachetools==5.0.0
# certifi==2021.10.8
# cffi==1.15.0
# charset-normalizer==2.0.12
# click==8.0.4
# cycler==0.11.0
# decorator==4.4.2
# filelock==3.6.0
# flatbuffers==2.0
# fonttools==4.29.1
# gast==0.5.3
# google-auth==2.6.0
# google-auth-oauthlib==0.4.6
# google-pasta==0.2.0
# grpcio==1.44.0
# h5py==3.6.0
# huggingface-hub==0.4.0
# idna==3.3
# imageio==2.16.0
# imageio-ffmpeg==0.4.5
# importlib-metadata==4.11.1
# joblib==1.1.0
# keras==2.8.0
# Keras-Preprocessing==1.1.2
# kiwisolver==1.3.2
# libclang==13.0.0
# librosa==0.9.1
# llvmlite==0.38.0
# Markdown==3.3.6
# matplotlib==3.5.1
# moviepy==1.0.3
# nltk==3.7
# noisereduce==2.0.0
# numba==0.55.1
# numpy==1.22.2
# oauthlib==3.2.0
# opencv-python==4.5.5.62
# opt-einsum==3.3.0
# packaging==21.3
# pika==1.2.0
# Pillow==9.0.1
# pooch==1.6.0
# proglog==0.1.9
# protobuf==3.19.4
# pyasn1==0.4.8
# pyasn1-modules==0.2.8
# pycparser==2.21
# pydub==0.25.1
# pyparsing==3.0.7
# python-dateutil==2.8.2
# python-dotenv==0.19.2
# PyYAML==6.0
# regex==2022.1.18
# requests==2.27.1
# requests-oauthlib==1.3.1
# resampy==0.2.2
# rsa==4.8
# sacremoses==0.0.47
# scikit-learn==1.0.2
# scipy==1.8.0
# sentence-transformers==2.2.0
# sentencepiece==0.1.96
# six==1.16.0
# sklearn==0.0
# SoundFile==0.10.3.post1
# SpeechRecognition==3.8.1
# tensorboard==2.8.0
# tensorboard-data-server==0.6.1
# tensorboard-plugin-wit==1.8.1
# tensorflow==2.8.0
# tensorflow-io-gcs-filesystem==0.24.0
# termcolor==1.1.0
# tf-estimator-nightly==2.8.0.dev2021122109
# threadpoolctl==3.1.0
# tokenizers==0.11.5
# torch==1.10.2
# torchvision==0.11.3
# tqdm==4.62.3
# transformers==4.16.2
# typing-extensions==4.1.1
# urllib3==1.26.8
# Werkzeug==2.0.3
# wrapt==1.13.3
# zipp==3.7.0

pydub>=0.25.1
noisereduce>=2.0.0
scipy>=1.7.1
numpy>=1.20.3
sklearn>=0.0
scikit-learn>=0.24.2
moviepy==1.0.3
SpeechRecognition
sentence-transformers
tensorflow-gpu
opencv-python-headless
tensorflow-cpu
pika==1.2.0
python-dotenv==0.19.2
